feat(desktop): share the workspace context menu with board cards
Extract the list row's right-click menu (Open, Copy Branch Name,
Pin/Unpin, Delete) plus its delete-dialog latch into a shared
V2WorkspaceContextMenu render-prop component and wrap board cards in it.
The render prop hands the same actions to inline affordances, so the
card's click-to-open drops its duplicated paywall gating and the row's
pin/trash buttons reuse one implementation. Archived ghost cards render
without the menu — no actions apply to a tombstone.
