+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspaceContextMenu/V2WorkspaceContextMenu.tsx
@@ -1,0 +1,191 @@
+import {
+	ContextMenu,
+	ContextMenuContent,
+	ContextMenuItem,
+	ContextMenuSeparator,
+	ContextMenuTrigger,
+} from "@superset/ui/context-menu";
+import { toast } from "@superset/ui/sonner";
+import { useNavigate } from "@tanstack/react-router";
+import { type ReactNode, useCallback, useState } from "react";
+import { LuArrowUpRight, LuGitBranch, LuTrash2 } from "react-icons/lu";
+import { RiPushpinFill, RiPushpinLine } from "react-icons/ri";
+import { GATED_FEATURES, usePaywall } from "renderer/components/Paywall";
+import { useCopyToClipboard } from "renderer/hooks/useCopyToClipboard";
+import { DashboardSidebarDeleteDialog } from "renderer/routes/_authenticated/_dashboard/components/DashboardSidebar/components/DashboardSidebarDeleteDialog";
+import { navigateToV2Workspace } from "renderer/routes/_authenticated/_dashboard/utils/workspace-navigation";
+import type { AccessibleV2Workspace } from "renderer/routes/_authenticated/_dashboard/v2-workspaces/hooks/useAccessibleV2Workspaces";
+import { useDashboardSidebarState } from "renderer/routes/_authenticated/hooks/useDashboardSidebarState";
+
+export interface V2WorkspaceActions {
+	/** Navigate to the workspace (paywall-gated for remote hosts). */
+	open: () => void;
+	addToSidebar: () => void;
+	removeFromSidebar: () => void;
+	openDeleteDialog: () => void;
+}
+
+interface V2WorkspaceContextMenuProps {
+	workspace: AccessibleV2Workspace;
+	/** Unpinning the current route's workspace is blocked. */
+	isCurrentRoute?: boolean;
+	/** Rendered as the context-menu trigger; receives the shared actions so
+	 * inline affordances (pin cell, trash button, card click) reuse them. */
+	children: (actions: V2WorkspaceActions) => ReactNode;
+}
+
+/**
+ * Right-click menu + delete-dialog wiring shared by the list rows and the
+ * board cards. Owns the workspace actions so both surfaces stay identical.
+ */
+export function V2WorkspaceContextMenu({
+	workspace,
+	isCurrentRoute = false,
+	children,
+}: V2WorkspaceContextMenuProps) {
+	const navigate = useNavigate();
+	const { gateFeature } = usePaywall();
+	const {
+		ensureWorkspaceInSidebar,
+		removeWorkspaceFromSidebar,
+		hideWorkspaceInSidebar,
+	} = useDashboardSidebarState();
+	const { copyToClipboard } = useCopyToClipboard();
+	const isMainWorkspace = workspace.type === "main";
+	const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
+	// Latched on first open so the dialog stays mounted through the destroy —
+	// a teardown failure re-opens it to offer force-delete.
+	const [deleteDialogMounted, setDeleteDialogMounted] = useState(false);
+
+	const open = useCallback(() => {
+		const go = () => navigateToV2Workspace(workspace.id, navigate);
+		if (workspace.hostType === "local-device") {
+			go();
+			return;
+		}
+		gateFeature(GATED_FEATURES.REMOTE_WORKSPACES, go);
+	}, [gateFeature, navigate, workspace.hostType, workspace.id]);
+
+	const addToSidebar = useCallback(() => {
+		const add = () =>
+			ensureWorkspaceInSidebar(workspace.id, workspace.projectId);
+		if (workspace.hostType === "local-device") {
+			add();
+			return;
+		}
+		gateFeature(GATED_FEATURES.REMOTE_WORKSPACES, add);
+	}, [
+		ensureWorkspaceInSidebar,
+		gateFeature,
+		workspace.hostType,
+		workspace.id,
+		workspace.projectId,
+	]);
+
+	const removeFromSidebar = useCallback(() => {
+		if (isCurrentRoute) return;
+		// Unpin directly (synchronous optimistic write) rather than routing
+		// through the intent store + RemoveFromSidebarMount effect, which adds
+		// an extra render cycle of latency. The list view is never a workspace
+		// route, so there's no active workspace to navigate away from.
+		//
+		// Always hide (keep the row with isHidden) rather than delete: the
+		// auto-add-local-workspaces hook treats a missing v2WorkspaceLocalState
+		// row as never-seen and would re-pin it. The tombstone row preserves the
+		// unpin intent.
+		hideWorkspaceInSidebar(workspace.id, workspace.projectId);
+	}, [
+		isCurrentRoute,
+		hideWorkspaceInSidebar,
+		workspace.id,
+		workspace.projectId,
+	]);
+
+	const handleCopyBranchName = useCallback(async () => {
+		try {
+			await copyToClipboard(workspace.branch);
+			toast.success("Branch name copied");
+		} catch (error) {
+			toast.error(
+				`Failed to copy branch name: ${error instanceof Error ? error.message : "Unknown error"}`,
+			);
+		}
+	}, [copyToClipboard, workspace.branch]);
+
+	const openDeleteDialog = useCallback(() => {
+		setDeleteDialogMounted(true);
+		setIsDeleteDialogOpen(true);
+	}, []);
+
+	const handleDeleted = useCallback(() => {
+		setDeleteDialogMounted(false);
+		removeWorkspaceFromSidebar(workspace.id);
+	}, [removeWorkspaceFromSidebar, workspace.id]);
+
+	return (
+		<>
+			<ContextMenu>
+				<ContextMenuTrigger asChild>
+					{children({
+						open,
+						addToSidebar,
+						removeFromSidebar,
+						openDeleteDialog,
+					})}
+				</ContextMenuTrigger>
+				<ContextMenuContent
+					onCloseAutoFocus={(event) => event.preventDefault()}
+				>
+					<ContextMenuItem onSelect={open}>
+						<LuArrowUpRight className="size-4" />
+						Open
+					</ContextMenuItem>
+					<ContextMenuItem onSelect={handleCopyBranchName}>
+						<LuGitBranch className="size-4" />
+						Copy Branch Name
+					</ContextMenuItem>
+					<ContextMenuSeparator />
+					{workspace.isInSidebar ? (
+						<ContextMenuItem
+							onSelect={removeFromSidebar}
+							disabled={isCurrentRoute}
+						>
+							<RiPushpinLine className="size-4" />
+							Unpin from Sidebar
+						</ContextMenuItem>
+					) : (
+						<ContextMenuItem onSelect={addToSidebar}>
+							<RiPushpinFill className="size-4" />
+							Pin to Sidebar
+						</ContextMenuItem>
+					)}
+					{!isMainWorkspace ? (
+						<>
+							<ContextMenuSeparator />
+							<ContextMenuItem
+								onSelect={openDeleteDialog}
+								className="text-destructive focus:text-destructive"
+							>
+								<LuTrash2 className="size-4 text-destructive" />
+								Delete
+							</ContextMenuItem>
+						</>
+					) : null}
+				</ContextMenuContent>
+			</ContextMenu>
+			{/* Mount the dialog (and its per-workspace live-query subscription)
+			    only once the user opened it — not idle for every row. The latch
+			    keeps it mounted through the destroy so a teardown-failure can
+			    re-open it to offer force-delete. */}
+			{!isMainWorkspace && deleteDialogMounted ? (
+				<DashboardSidebarDeleteDialog
+					workspaceId={workspace.id}
+					workspaceName={workspace.name || workspace.branch}
+					open={isDeleteDialogOpen}
+					onOpenChange={setIsDeleteDialogOpen}
+					onDeleted={handleDeleted}
+				/>
+			) : null}
+		</>
+	);
+}
