+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspaces/components/V2WorkspaceContextMenu/index.ts
@@ -1,0 +1,4 @@
+export {
+	type V2WorkspaceActions,
+	V2WorkspaceContextMenu,
+} from "./V2WorkspaceContextMenu";
